move storage wrapper and interface to extension
