+++ apps/extension/src/ctx/spend-key.ts
@@ -1,7 +1,7 @@
 import { Code, ConnectError } from '@connectrpc/connect';
 import { Key } from '@penumbra-zone/crypto-web/encryption';
-import { localExtStorage } from '@penumbra-zone/storage/chrome/local';
-import { sessionExtStorage } from '@penumbra-zone/storage/chrome/session';
+import { localExtStorage } from '../storage/local';
+import { sessionExtStorage } from '../storage/session';
 import { Box } from '@penumbra-zone/types/box';
 import { generateSpendKey } from '@penumbra-zone/wasm/keys';
 
@@ -11,7 +11,7 @@ export const getSpendKey = async () => {
   const passKey = await Key.fromJson(passKeyJson);
 
   const wallet0 = (await localExtStorage.get('wallets'))[0];
-  if (!wallet0) throw new ConnectError('No wallet found', Code.FailedPrecondition);
+  if (!wallet0) throw new ConnectError('No wallet found');
 
   const seedBox = Box.fromJson(wallet0.custody.encryptedSeedPhrase);
   const seedPhrase = await passKey.unseal(seedBox);
