+++ apps/extension/src/ctx/wallet-id.ts
@@ -1,6 +1,6 @@
 import { WalletId } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/keys/v1/keys_pb';
 import { Code, ConnectError } from '@connectrpc/connect';
-import { localExtStorage } from '@penumbra-zone/storage/chrome/local';
+import { localExtStorage } from '../storage/local';
 
 export const getWalletId = async () => {
   const wallet0 = (await localExtStorage.get('wallets'))[0];
