+++ apps/extension/src/routes/popup/home/index.tsx
@@ -2,7 +2,7 @@ import { SelectAccount } from '@penumbra-zone/ui/components/ui/select-account';
 import { IndexHeader } from './index-header';
 import { useStore } from '../../../state';
 import { BlockSync } from './block-sync';
-import { localExtStorage } from '@penumbra-zone/storage/chrome/local';
+import { localExtStorage } from '../../../storage/local';
 import { getActiveWallet } from '../../../state/wallets';
 import { needsLogin } from '../popup-needs';
 import {
