+++ apps/extension/src/routes/popup/settings/settings-connected-sites/known-sites-group.tsx
@@ -1,4 +1,4 @@
-import { OriginRecord } from '@penumbra-zone/storage/chrome/types';
+import { OriginRecord } from '../../../../storage/types';
 import { useStore } from '../../../../state';
 import { KnownSite } from './known-site';
