+++ apps/extension/src/state/connected-sites.test.ts
@@ -1,12 +1,9 @@
 import { beforeEach, describe, expect, test } from 'vitest';
 import { create, StoreApi, UseBoundStore } from 'zustand';
 import { AllSlices, initializeStore } from '.';
-import { LocalStorageState, OriginRecord } from '@penumbra-zone/storage/chrome/types';
-import { ExtensionStorage } from '@penumbra-zone/storage/chrome/base';
-import {
-  mockLocalExtStorage,
-  mockSessionExtStorage,
-} from '@penumbra-zone/storage/chrome/test-utils/mock';
+import { LocalStorageState, OriginRecord } from '../storage/types';
+import { ExtensionStorage } from '../storage/base';
+import { mockLocalExtStorage, mockSessionExtStorage } from '../storage/mock';
 import { UserChoice } from '@penumbra-zone/types/user-choice';
 import { allSitesFilteredOutSelector } from './connected-sites';
 import { localTestDefaults } from '../utils/test-constants';
