+++ apps/extension/src/state/default-frontend.test.ts
@@ -1,12 +1,9 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 import { create, StoreApi, UseBoundStore } from 'zustand';
 import { AllSlices, initializeStore } from '.';
-import { LocalStorageState } from '@penumbra-zone/storage/chrome/types';
-import { ExtensionStorage } from '@penumbra-zone/storage/chrome/base';
-import {
-  mockLocalExtStorage,
-  mockSessionExtStorage,
-} from '@penumbra-zone/storage/chrome/test-utils/mock';
+import { LocalStorageState } from '../storage/types';
+import { ExtensionStorage } from '../storage/base';
+import { mockLocalExtStorage, mockSessionExtStorage } from '../storage/mock';
 
 describe('Default Frontend Slice', () => {
   let useStore: UseBoundStore<StoreApi<AllSlices>>;
