+++ apps/extension/src/state/index.ts
@@ -5,10 +5,10 @@ import { customPersist } from './persist';
 import { createPasswordSlice, PasswordSlice } from './password';
 import { createSeedPhraseSlice, SeedPhraseSlice } from './seed-phrase';
 import { createNetworkSlice, NetworkSlice } from './network';
-import { localExtStorage } from '@penumbra-zone/storage/chrome/local';
-import { LocalStorageState } from '@penumbra-zone/storage/chrome/types';
-import { sessionExtStorage, SessionStorageState } from '@penumbra-zone/storage/chrome/session';
-import { ExtensionStorage } from '@penumbra-zone/storage/chrome/base';
+import { localExtStorage } from '../storage/local';
+import { LocalStorageState } from '../storage/types';
+import { sessionExtStorage, SessionStorageState } from '../storage/session';
+import { ExtensionStorage } from '../storage/base';
 import { createTxApprovalSlice, TxApprovalSlice } from './tx-approval';
 import { createOriginApprovalSlice, OriginApprovalSlice } from './origin-approval';
 import { ConnectedSitesSlice, createConnectedSitesSlice } from './connected-sites';
