+++ apps/extension/src/state/network.test.ts
@@ -1,12 +1,9 @@
 import { beforeEach, describe, expect, test } from 'vitest';
 import { create, StoreApi, UseBoundStore } from 'zustand';
 import { AllSlices, initializeStore } from '.';
-import { LocalStorageState } from '@penumbra-zone/storage/chrome/types';
-import {
-  mockLocalExtStorage,
-  mockSessionExtStorage,
-} from '@penumbra-zone/storage/chrome/test-utils/mock';
-import { ExtensionStorage } from '@penumbra-zone/storage/chrome/base';
+import { LocalStorageState } from '../storage/types';
+import { mockLocalExtStorage, mockSessionExtStorage } from '../storage/mock';
+import { ExtensionStorage } from '../storage/base';
 
 describe('Network Slice', () => {
   let useStore: UseBoundStore<StoreApi<AllSlices>>;
