+++ apps/extension/src/state/seed-phrase/import.test.ts
@@ -2,10 +2,7 @@ import { create, StoreApi, UseBoundStore } from 'zustand';
 import { AllSlices, initializeStore } from '..';
 import { beforeEach, describe, expect, test } from 'vitest';
 import { SeedPhraseLength } from '@penumbra-zone/crypto-web/mnemonic';
-import {
-  mockLocalExtStorage,
-  mockSessionExtStorage,
-} from '@penumbra-zone/storage/chrome/test-utils/mock';
+import { mockLocalExtStorage, mockSessionExtStorage } from '../../storage/mock';
 
 describe('Import Slice', () => {
   let useStore: UseBoundStore<StoreApi<AllSlices>>;
