+++ apps/extension/src/state/wallets.ts
@@ -1,8 +1,8 @@
 import { AllSlices, SliceCreator } from '.';
 import { generateSpendKey, getFullViewingKey, getWalletId } from '@penumbra-zone/wasm/keys';
 import { Key } from '@penumbra-zone/crypto-web/encryption';
-import { ExtensionStorage } from '@penumbra-zone/storage/chrome/base';
-import { LocalStorageState } from '@penumbra-zone/storage/chrome/types';
+import { ExtensionStorage } from '../storage/base';
+import { LocalStorageState } from '../storage/types';
 import {
   FullViewingKey,
   SpendKey,
