+++ apps/extension/src/storage/base.ts
