+++ apps/extension/src/storage/mock.ts
@@ -1,7 +1,7 @@
-import { ExtensionStorage, IStorage } from '../base';
-import { localDefaults } from '../local';
-import { sessionDefaults, SessionStorageState } from '../session';
-import { LocalStorageState } from '../types';
+import { ExtensionStorage, IStorage } from './base';
+import { localDefaults } from './local';
+import { sessionDefaults, SessionStorageState } from './session';
+import { LocalStorageState } from './types';
 
 // Helpful for testing interactions with session & local storage
 export class MockStorageArea implements IStorage {
