+++ apps/extension/src/storage/session.ts
